backend/nosylos_api/requirements.txt
asgiref==3.5.2
backports.zoneinfo==0.2.1;python_version<"3.9"
boto3==1.26.27
botocore==1.29.27
certifi==2022.12.7
cffi==1.15.1
cfgv==3.3.1
charset-normalizer==3.0.1
colorama==0.4.6
cryptography==38.0.1
distlib==0.3.6
Django==5.2
django-cors-headers==3.13.0
django-extensions==3.2.1
django-filter==22.1
django-request==1.6.3
django-ses==3.2.2
djangorestframework==3.14.0
djangorestframework-camel-case==1.4.2
djangorestframework-simplejwt==5.2.1
filelock==3.8.0
identify==2.5.6
idna==3.4
importlib-metadata==5.0.0
iniconfig==2.0.0
jmespath==1.0.1
Markdown==3.4.1
nodeenv==1.7.0
packaging==23.2
platformdirs==2.5.2
pluggy==1.3.0
pre-commit==2.20.0
psycopg2-binary==2.9.5
pycparser==2.21
PyJWT==2.5.0
python-dateutil==2.8.2
pytz==2022.4
PyYAML==6.0
requests==2.28.2
s3transfer==0.6.0
six==1.16.0
sqlparse==0.4.3
stripe==7.6.0
toml==0.10.2
typing_extensions==4.8.0
tzdata==2022.4
urllib3==1.26.13
virtualenv==20.16.5
zipp==3.8.1
